New component for init keycloak
