+++ simpipe-frontend/src/routes/index.svelte
@@ -1,38 +1,20 @@
 <script>
-	import { GraphQLClient } from 'graphql-request';
 	import { all_simulations_query } from '../queries/all_simulations.svelte';
-	import Keycloak from 'keycloak-js';
 	import { graphQLClient } from '../stores/stores';
 	import { get } from 'svelte/store';
 	import Modal from 'svelte-simple-modal';
 	import CreateSimulationButton from './modals/create_simulation_button.svelte';
 	import { simulations_list } from '../stores/stores';
-	export let simulations;
+	import { init_keycloak } from './keycloak.svelte';
+
+	let simulations;	
 	let loading = async () => {
 		try {
-			let keycloak = new Keycloak({
-					url: 'https://datacloud-auth.euprojects.net/auth', // todo - add env var
-					realm: 'user-authentication',
-					clientId: 'sim-pipe-web', // todo - add env var
-					'enable-cors': 'true'
-				});
-				await keycloak.init({ onLoad: 'login-required', flow: 'implicit' }); //  Implicit flow. 
-				console.log('User authenticated');
-				const requestHeaders = {
-					authorization: 'Bearer '  + keycloak.token,
-					mode: 'cors',
-				}
-				// todo - add confi
-				graphQLClient.set(new GraphQLClient('http://localhost:9000/graphql', {
-					headers: requestHeaders
-				}));
-				simulations = await get(graphQLClient).request(all_simulations_query);
-		}		
-		catch {
-	
-		}
-
-	}
+			await init_keycloak();
+			simulations = await get(graphQLClient).request(all_simulations_query);
+			// $: $simulations_list = await get(graphQLClient).request(all_simulations_query);
+		} catch {}
+	};
 	loading();
 	$: $simulations_list = simulations;
 </script>
@@ -42,8 +24,8 @@
 	<ul class="max_width">
 		<li class="table-header">Simulations</li>
 		<br />
-		{#if !simulations} 
-			<p>Loading simulations... </p>
+		{#if !simulations}
+			<p>Loading simulations...</p>
 		{:else}
 			{#each $simulations_list.All_Simulations.simulations as simulation}
 				<li class="pointer row">
